Add Simon animations and a controller to control her
- Add most of Simon animations (jump, attack, duck...)
- Add Simon controller to control simon character
- Split all gameobjects in a seperate categories (Entity, Boundary, Trigger..)
- Reset animation to the starting frame when changing animation
- ChracterStatsReader now return a generic dictionary of object properites instead of hardcode properties (more flexibility)
- Increase tile size from 32x32 to 64x64 (too much tile to draw seem to eat a lot of CPU cycles)
- Revert to using event in input system. Comparing current state and prev state of keyboard
  does not works because of the queueing nature of gameloop, event-based input is more responsive
- Fix wrong game object positions in tiled map file

+++ src/GameCuaTao/Direct2DGame/Graphics/Viewport.h
@@ -5,8 +5,8 @@
 
 struct Viewport
 {
-	int x;
-	int y;
+	float x;
+	float y;
 	int width;
 	int height;
 
@@ -24,6 +24,6 @@ struct Viewport
 
 	Rect Bounds()
 	{
-		return Rect(x, y, width, height);
+		return Rect((int)x, (int)y, width, height);
 	}
 };
